requirements.txt
# Web框架
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# LangChain & LangGraph
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langgraph==0.2.0

# 搜索和解析
tavily-python==0.5.0
firecrawl-py==1.5.0
beautifulsoup4==4.12.0
requests==2.32.0
aiohttp==3.10.0

# 向量存储和NLP
chromadb==0.5.0
sentence-transformers==3.0.0
numpy==1.26.0
tiktoken==0.7.0

# 数据验证和工具
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
httpx==0.27.0

# 异步和任务队列
celery==5.4.0
redis==5.0.0

# 日志和监控
structlog==24.4.0
prometheus-client==0.21.0

# 测试
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# 开发工具
black==24.8.0
isort==5.13.0
mypy==1.11.0
